+++ fourPolar-core/src/main/java/fr/fresnel/fourPolar/core/physics/dipole/OrientationVector.java
@@ -6,24 +6,24 @@ import fr.fresnel.fourPolar.core.exceptions.physics.dipole.OrientationAngleOutOf
  * Models the orientation angles calculated using the four polar method.
  */
 public class OrientationVector implements IOrientationVector {
-    final private float _rho;
-    final private float _delta;
-    final private float _eta;
+    final private double _rho;
+    final private double _delta;
+    final private double _eta;
 
     /**
      * Maximum possible value for the rho
      */
-    public final static float MAX_Rho = (float) Math.PI;
+    public final static double MAX_Rho = Math.PI;
 
     /**
      * Maximum possible value for the delta
      */
-    public final static float MAX_Delta = (float) Math.PI;
+    public final static double MAX_Delta = Math.PI;
 
     /**
      * Maximum possible value for the delta
      */
-    public final static float MAX_Eta = (float) Math.PI / 2;
+    public final static double MAX_Eta = Math.PI / 2;
 
     /**
      * Models the orientation angles calculated using the four polar method. Note
@@ -34,11 +34,11 @@ public class OrientationVector implements IOrientationVector {
      * @param delta : delta angle in radian, ranges from 0 to pi.
      * @param eta   : eta angle in radian, ranges from 0 to pi/2.
      */
-    public OrientationVector(float rho, float delta, float eta) throws OrientationAngleOutOfRange {
-        if (Float.isNaN(rho) || Float.isNaN(delta) || Float.isNaN(eta)) {
-            _rho = Float.NaN;
-            _delta = Float.NaN;
-            _eta = Float.NaN;
+    public OrientationVector(double rho, double delta, double eta) throws OrientationAngleOutOfRange {
+        if (Double.isNaN(rho) || Double.isNaN(delta) || Double.isNaN(eta)) {
+            _rho = Double.NaN;
+            _delta = Double.NaN;
+            _eta = Double.NaN;
         } else {
             _checkRho(rho);
             _rho = rho;
@@ -49,12 +49,12 @@ public class OrientationVector implements IOrientationVector {
             _checkEta(eta);
             _eta = eta;
         }
-
+ 
     }
 
     @Override
-    public float getAngle(OrientationAngle angle) {
-        float orientation = 0;
+    public double getAngle(OrientationAngle angle) {
+        double orientation = 0;
         switch (angle) {
             case rho:
                 orientation = _rho;
@@ -76,23 +76,23 @@ public class OrientationVector implements IOrientationVector {
     }
 
     @Override
-    public float getAngleInDegree(OrientationAngle angle) {
-        return (float) Math.toDegrees(this.getAngle(angle));
+    public double getAngleInDegree(OrientationAngle angle) {
+        return Math.toDegrees(this.getAngle(angle));
     }
 
-    private void _checkRho(float value) throws OrientationAngleOutOfRange {
+    private void _checkRho(double value) throws OrientationAngleOutOfRange {
         if (value < 0 || value >= MAX_Rho) {
             throw new OrientationAngleOutOfRange("Rho is out of [0, pi) range");
         }
     }
 
-    private void _checkDelta(float value) throws OrientationAngleOutOfRange {
+    private void _checkDelta(double value) throws OrientationAngleOutOfRange {
         if (value < 0 || value > MAX_Delta) {
             throw new OrientationAngleOutOfRange("Delta is out of [0, pi] range");
         }
     }
 
-    private void _checkEta(float value) throws OrientationAngleOutOfRange {
+    private void _checkEta(double value) throws OrientationAngleOutOfRange {
         if (value < 0 || value >= MAX_Eta) {
             throw new OrientationAngleOutOfRange("Eta is out of [0, pi/2) range");
         }
