+++ fourPolar-core/src/test/java/fr/fresnel/fourPolar/core/image/orientation/OrientationVectorIteratorTest.java
@@ -68,14 +68,14 @@ public class OrientationVectorIteratorTest {
             Float32 angle = baseSetCursor.next().value();
 
             equals &= _checkPrecision(angle.get(), orientationVector.getAngle(OrientationAngle.rho), 1e-6f);
-            equals &= _checkPrecision(angle.get(), orientationVector.getAngle(OrientationAngle.delta) - (float)Math.PI/180, 1e-6f);
-            equals &= _checkPrecision(angle.get(), orientationVector.getAngle(OrientationAngle.eta) - (float)Math.PI/90, 1e-6f);
+            equals &= _checkPrecision(angle.get(), orientationVector.getAngle(OrientationAngle.delta) - Math.PI/180, 1e-6f);
+            equals &= _checkPrecision(angle.get(), orientationVector.getAngle(OrientationAngle.eta) - Math.PI/90, 1e-6f);
         }
 
         assertTrue(equals);
     }
 
-    private static boolean _checkPrecision(float val1, float val2, float error) {
+    private static boolean _checkPrecision(double val1, double val2, double error) {
         return Math.abs(val1 - val2) < error;        
     }
